+++ test/cli/test_commands.py
@@ -2347,8 +2347,8 @@ def test_warning_on_all_not_found(cli, openapi3_schema_url, openapi3_base_url):
     # Then the output should contain a warning about it
     assert result.exit_code == ExitCode.OK, result.stdout
     assert (
-        "WARNING: All API responses have a 404 status code. "
-        "Did you specify the proper API location?" in strip_style_win32(result.stdout)
+        "`GET /api/v4/success` returned only 4xx responses during unit tests. Check base URL or adjust data generation settings"
+        in strip_style_win32(result.stdout)
     )